rag-notes-helper/requirements.txt
faiss-cpu>=1.8
sentence-transformers>=3.0
pypdf>=4.0
fastapi>=0.110
uvicorn>=0.30
numpy>=1.26
pytest>=8.0
